requirements.txt
requests>=2.31.0
regex>=2023.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
